Add files via upload
Data cleaning is the process of identifying, correcting, and removing errors, inconsistencies, and inaccuracies in datasets. It involves a series of steps such as data validation, data transformation, data normalization, and data enrichment to ensure that the data is accurate, complete, and consistent. The purpose of data cleaning is to improve the quality and reliability of data, making it more suitable for analysis, reporting, and decision-making. Data cleaning is a critical step in data preprocessing and is often performed before data analysis or machine learning algorithms are applied to the data.